Updated runner to use context and file parser
Necessary to leverage new parser now that the writer has been removed.
Additionally, the object API has been updated to be an interactor.

A context has been added to encapsulate the multiple attributes
necessary for the runner to execute properly. The `.for` method was
added to make it easier to splat multiple attributes if desired.
This'll help with downstream objects who may or may not want to
construct an context first.