Updated comments and writer to use main formatter
Necessary to leverage the new formatter now that the old implementation
has been removed. Additionally, method names were updated as a minor
improvement in readability since *insert* and *remove* are better
antonyms of each other.

This object was moved into the parsers namespace since it deals with
the formatting and comment unions and/or intersections.