Updated CLI to use insert instead of add option
Due to recent changes to the runner, the CLI needed to be updated to
pass in the proper arguments to the runner. This includes removing the
`--add` option in favor of the `--insert` option to better mimic the
internal API.

Associated specs were simplified since the heavier testing is handled
in the `Runner` specs.

This is a breaking change as there is no backwards compatibility
support.