Added CLI shell
Provides the command line interface for this gem and replaces the work
originally provided by the recently deleted CLI object.

The Ruby 2.7.0 experimental feature warning is disabled in order to not
display warnings or confuse people. The warning is left enabled for
development purposes, though. Disablement of the experimental feature
warning will need to be removed in the future once the Ruby core team
moves this support out of experimental status.