Fixed CLI action optional positional parameters
Reek was updated to ignore the control parameter issues because this is a side effect of using `OptionParser` since it doesn't allow you to use an optional positional parameter (it's essentially ignored).

Use of struct refinements isn't necessary in this situation so is safe to remove which simplifies the implementation further.